+++ tests/mock_values/restful/wallet/getApiKeyPermission.ts
@@ -11,5 +11,5 @@ export const mockResponse: RestWalletTypes.getApiKeyPermissionResponse = {
     'enableFutures': false,
     'enableMargin': false,
     'enableSpotAndMarginTrading': false,
-    'tradingAuthorityExpirationTime': 1628985600000
+    'enablePortfolioMarginTrading': false
 };